+++ lib/core/widgets/components/coffee_card.dart
