+++ lib/core/widgets/components/coffee_card_placeholder.dart
