+++ lib/core/widgets/components/error_section.dart
@@ -1,5 +1,5 @@
 import 'package:coffeecard/core/strings.dart';
-import 'package:coffeecard/core/widgets/components/tickets/rounded_button.dart';
+import 'package:coffeecard/core/widgets/components/rounded_button.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
