+++ lib/core/widgets/components/single_coffee_card.dart
