+++ lib/features/product/presentation/widgets/buy_ticket_bottom_modal_sheet.dart
@@ -1,8 +1,8 @@
 import 'package:coffeecard/core/strings.dart';
 import 'package:coffeecard/core/styles/app_colors.dart';
 import 'package:coffeecard/core/styles/app_text_styles.dart';
-import 'package:coffeecard/core/widgets/components/tickets/bottom_modal_sheet_helper.dart';
-import 'package:coffeecard/core/widgets/components/tickets/rounded_button.dart';
+import 'package:coffeecard/core/widgets/components/bottom_modal_sheet_helper.dart';
+import 'package:coffeecard/core/widgets/components/rounded_button.dart';
 import 'package:coffeecard/features/product/domain/entities/product.dart';
 import 'package:coffeecard/features/purchase/domain/entities/internal_payment_type.dart';
 import 'package:coffeecard/features/purchase/domain/entities/payment.dart';
