+++ lib/features/ticket/presentation/widgets/swipe_ticket_confirm.dart
@@ -3,9 +3,9 @@ import 'dart:async';
 import 'package:coffeecard/core/strings.dart';
 import 'package:coffeecard/core/styles/app_colors.dart';
 import 'package:coffeecard/core/styles/app_text_styles.dart';
+import 'package:coffeecard/core/widgets/components/bottom_modal_sheet_helper.dart';
 import 'package:coffeecard/core/widgets/components/card.dart';
-import 'package:coffeecard/core/widgets/components/tickets/bottom_modal_sheet_helper.dart';
-import 'package:coffeecard/core/widgets/components/tickets/slide_to_act.dart';
+import 'package:coffeecard/core/widgets/components/slide_action.dart';
 import 'package:coffeecard/features/receipt/presentation/cubit/receipt_cubit.dart';
 import 'package:coffeecard/features/ticket/presentation/cubit/tickets_cubit.dart';
 import 'package:flutter/material.dart';
