+++ test/core/widgets/components/tickets/coffee_card_test.dart
@@ -1,4 +1,4 @@
-import 'package:coffeecard/core/widgets/components/tickets/coffee_card.dart';
+import 'package:coffeecard/core/widgets/components/coffee_card.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
